+++ mod/twitch/src/main/java/tv/twitch/android/feature/theatre/common/StreamSettingsViewDelegate.java
@@ -1,0 +1,9 @@
+package tv.twitch.android.feature.theatre.common;
+
+import androidx.fragment.app.FragmentActivity;
+
+import tv.twitch.android.core.mvp.viewdelegate.BaseViewDelegate;
+
+public class StreamSettingsViewDelegate extends BaseViewDelegate {
+    public FragmentActivity activity;
+}